refactor form.js
